+++ x/protocolpool/keeper/keeper.go
@@ -1,0 +1,69 @@
+package keeper
+
+import (
+	"context"
+	"fmt"
+
+	storetypes "cosmossdk.io/core/store"
+	errorsmod "cosmossdk.io/errors"
+	"cosmossdk.io/log"
+	"cosmossdk.io/x/protocolpool/types"
+
+	"github.com/cosmos/cosmos-sdk/codec"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+)
+
+type Keeper struct {
+	storeService storetypes.KVStoreService
+	authKeeper   types.AccountKeeper
+	bankKeeper   types.BankKeeper
+
+	authority string
+}
+
+func NewKeeper(cdc codec.BinaryCodec, storeService storetypes.KVStoreService,
+	ak types.AccountKeeper, bk types.BankKeeper, authority string,
+) Keeper {
+	// ensure pool module account is set
+	if addr := ak.GetModuleAddress(types.ModuleName); addr == nil {
+		panic(fmt.Sprintf("%s module account has not been set", types.ModuleName))
+	}
+	return Keeper{
+		storeService: storeService,
+		authKeeper:   ak,
+		bankKeeper:   bk,
+		authority:    authority,
+	}
+}
+
+// GetAuthority returns the x/protocolpool module's authority.
+func (k Keeper) GetAuthority() string {
+	return k.authority
+}
+
+// Logger returns a module-specific logger.
+func (k Keeper) Logger(ctx context.Context) log.Logger {
+	sdkCtx := sdk.UnwrapSDKContext(ctx)
+	return sdkCtx.Logger().With(log.ModuleKey, "x/"+types.ModuleName)
+}
+
+// FundCommunityPool allows an account to directly fund the community fund pool.
+func (k Keeper) FundCommunityPool(ctx context.Context, amount sdk.Coins, sender sdk.AccAddress) error {
+	return k.bankKeeper.SendCoinsFromAccountToModule(ctx, sender, types.ModuleName, amount)
+}
+
+// DistributeFromFeePool distributes funds from the protocolpool module account to
+// a receiver address.
+func (k Keeper) DistributeFromFeePool(ctx context.Context, amount sdk.Coins, receiveAddr sdk.AccAddress) error {
+	return k.bankKeeper.SendCoinsFromModuleToAccount(ctx, types.ModuleName, receiveAddr, amount)
+}
+
+// GetCommunityPool get the community pool balance.
+func (k Keeper) GetCommunityPool(ctx context.Context) (sdk.Coins, error) {
+	moduleAccount := k.authKeeper.GetModuleAccount(ctx, types.ModuleName)
+	if moduleAccount == nil {
+		return nil, errorsmod.Wrapf(sdkerrors.ErrUnknownAddress, "module account %s does not exist", moduleAccount)
+	}
+	return k.bankKeeper.GetAllBalances(ctx, moduleAccount.GetAddress()), nil
+}